+++ pom.xml
@@ -72,6 +72,13 @@
       <version>3.2.1</version>
       <scope>compile</scope>
     </dependency>
+
+    <!-- https://mvnrepository.com/artifact/org.apache.spark/spark-mllib -->
+    <dependency>
+      <groupId>org.apache.spark</groupId>
+      <artifactId>spark-mllib_2.13</artifactId>
+      <version>3.3.1</version>
+    </dependency>
   </dependencies>
 
   <build>
